+++ R/00base.R
@@ -1,2 +1,1 @@
 BASEURL <- "http://www.ebi.ac.uk/proteins/api/"
-PROTEINSURL <- paste0(BASEURL, "proteins?")