+++ NEWS.md
@@ -1,1 +1,3 @@
 # Changes in version 0.1.0
+
+- included first test function for Proteins <2016-12-27 Tue>